+++ project/src/App.tsx
@@ -4,26 +4,32 @@ import { Switch, BrowserRouter as Router, Route } from 'react-router-dom';
 import store from './store/index';
 
 import BoardView from './views/BoardView';
-import MainView from './views/MainView';
+
 import DashboardView from './views/DashboardView';
-import LoginView from './views/LoginView';
+
+import SettingsView from './views/SettingsView';
+import RaportsView from './views/RaportsView';
+import ScheduleView from './views/ScheduleView';
 
 function App() {
     return (
         <Provider store={store}>
             <Router>
                 <Switch>
-                    <Route path="/xxxx">
-                        <MainView></MainView>
-                    </Route>
                     <Route exact path="/">
                         <BoardView></BoardView>
                     </Route>
                     <Route path="/Dashboards">
                         <DashboardView></DashboardView>
                     </Route>
-                    <Route path="/LogIn">
-                        <LoginView></LoginView>
+                    <Route path="/Settings">
+                        <SettingsView></SettingsView>
+                    </Route>
+                    <Route path="/Raports">
+                        <RaportsView></RaportsView>
+                    </Route>
+                    <Route path="/Schedule">
+                        <ScheduleView></ScheduleView>
                     </Route>
                 </Switch>
             </Router>