+++ project/src/components/organisms/Sidebars/VerticalSidebar.tsx
@@ -6,8 +6,8 @@ import allColors from '../../../constants/allColors';
 const Sidebar = styled.div`
     display: flex;
     flex-direction: column;
-    justify-content: flex-start;
-    height: 80vh;
+    justify-content: space-between;
+    height: 90vh;
     width: 10vw;
     position: absolute;
     top: 0px;
@@ -17,25 +17,54 @@ const Sidebar = styled.div`
     margin: 0;
 `;
 
+const TopWrapper = styled.div`
+    height: 70vh;
+`;
+
+const BottomWrapper = styled.div`
+    justify-self: center;
+    margin-bottom: 10px;
+`;
+
 const VerticalSidebar = () => {
     return (
         <Sidebar>
-            <VerticalSidebarItem
-                path="/"
-                color={allColors.lightBlue}
-                iconPath={require('../../../assets/SVGs/dashboardGrey.svg')}
-                iconPathActive={require('../../../assets/SVGs/dashboard.svg')}
-            >
-                Boards
-            </VerticalSidebarItem>
-            <VerticalSidebarItem
-                path="/LogIn"
-                color={allColors.green}
-                iconPath={require('../../../assets/SVGs/calendar.svg')}
-                iconPathActive={require('../../../assets/SVGs/calendarGreen.svg')}
-            >
-                Schedule
-            </VerticalSidebarItem>
+            <TopWrapper>
+                <VerticalSidebarItem
+                    path="/"
+                    color={allColors.lightBlue}
+                    iconPath={require('../../../assets/SVGs/dashboardGrey.svg')}
+                    iconPathActive={require('../../../assets/SVGs/dashboard.svg')}
+                >
+                    Boards
+                </VerticalSidebarItem>
+                <VerticalSidebarItem
+                    path="/Schedule"
+                    color={allColors.green}
+                    iconPath={require('../../../assets/SVGs/calendar.svg')}
+                    iconPathActive={require('../../../assets/SVGs/calendarGreen.svg')}
+                >
+                    Schedule
+                </VerticalSidebarItem>
+                <VerticalSidebarItem
+                    path="/Raports"
+                    color={allColors.pink}
+                    iconPath={require('../../../assets/SVGs/statistics.svg')}
+                    iconPathActive={require('../../../assets/SVGs/statisticsPink.svg')}
+                >
+                    Raports
+                </VerticalSidebarItem>
+            </TopWrapper>
+            <BottomWrapper>
+                <VerticalSidebarItem
+                    path="/Settings"
+                    color={allColors.orange}
+                    iconPath={require('../../../assets/SVGs/settings.svg')}
+                    iconPathActive={require('../../../assets/SVGs/settingsOrange.svg')}
+                >
+                    Settings
+                </VerticalSidebarItem>
+            </BottomWrapper>
         </Sidebar>
     );
 };