+++ project/src/constants/allColors.ts
@@ -5,9 +5,11 @@ export default {
     light: '#DAECF7',
     dark: '#2D2A32',
     pinkish: '#F7567C',
-    grey: 'rgb(216,216,228)',
+    grey: 'rgb(201,206,213)',
     borderLightBlue: 'rgb(153,197,251)',
     lightBlue: 'rgb(146,170,251)',
     mediumBlue: 'rgb(132,144,219)',
     green: 'rgb(127,212,186)',
+    pink: 'rgb(224,118,142)',
+    orange: 'rgb(255,182,79)',
 };